add NMEA data extraction
this software state can validate if a received sentence is a
GPGGA-sentence and if it contains valid fix data. if so, data is
extracted and written to char arrays of fixed lengths. these can be
placed inside the APRS string and therefore modify the sent string -
therefore keeping down memory consumption

pc unit tests are included