clean up main, finish HW init
HW is now initialized properly. si4060-reset is done in main, which may
change in the future.

state up to now is: si4060 is responding to all commands with correct
CTS, except after "power up" ... it remains unresponsive afterwards,
only a full reset helps ...

chip id (part_info) can be read correctly, and a correct read is
indicated by turning of one LED on the fraunchpad.