enable watchdog, temp format fix, adc enabled
watchdog: as it was observed that the Si4060 locks up occasionally
(after power up, as known..) caused by clock issues, and in case the
software contains error for flight - the watchdog was enabled with about
11 seconds reset interval. this is more than enough for all working
situations. tested and verified ok!

temp format fix: the temperature is now output as signed and correctly
converted to two character decimal. tested only for ~+30C - tests in
climate chamber should tell us more :)

adc: the get_battery_voltage function was finished by enabling the ADC.
it was tested with 0V, 3V and mid scale voltage - seems to fit nicely