integrate u-Blox GNSS module
with the change from the Linx GPS to the u-Blox modules, some software
changes were neccesary. it requires much more configuration, therefore
the GPS init routines were factored out into gps.c

as it also utilizes different talker IDs, GPGGA was removed and changed
to a generic *GGA

at last, the new powersaving methods were implemented, but are yet to be
tested. this is a state where the power save/cyclic tracking mode seemed
not to be activated - but the uBlox ACKs the comand.