revert changes to use the uBlox TP as clk source
it soon got clear that the timepulse from the uBlox is not suitable as a
reference clock to the Si4060.

it often locks up, hanging the software, etc. a timeout feature for
failed start attempts was added, but is outcommented now, as it only
hides problems. if this should be handled for real, a proper
reset/reinit framework would be needed.