transmit sine wave in FM FSK
this commit makes the Si4060 transmit a 1200 Hz sine wave, frequency
modulated with +-3kHz deviation. works well so far, as soon as the
si4060 started to power up in the first place. afterwards, SPI
communication is rock solid, no hangs or crashes observed. the Si4060
seems to take longer to update the offset-register every second or so,
maybe because of internal interrupts?