APRS modulation filtering via Si4060 GFSK FIR-Filter
The Si4060 houses a FIR-Filter that can be used in GFSK mode for
spectral shaping of output data. This filter has 17 Taps which are
programmed with low pass filter coefficients. The filter is running at
10*4400 Hz, the input is sampled at 4400 Hz. The current filter gives
acceptable suppression of harmonics and offers pre-emphasis (6dB between
1200Hz and 2200 Hz).

Still has introduced a small reboot-bug, which has to be found & fixed.